Rename disk_pusher.py to stringer.py
